+++ src/main/java/insane96mcp/iguanatweaksreborn/module/combat/Knockback.java
@@ -121,7 +121,7 @@ public class Knockback extends Feature {
 
 	public static float getKnockbackMultiplier(ItemStack stack) {
 		float multiplier = 1f;
-		for (ItemDefinition itemDefinition : ItemDefinitionsReloadListener.getDefinitions()) {
+		for (ItemDefinition itemDefinition : ItemDefinitionsReloadListener.DEFINITIONS) {
 			if (itemDefinition.knockbackMultiplier() != null && itemDefinition.item().matchesItem(stack))
 				multiplier = itemDefinition.knockbackMultiplier().floatValue();
 		}