+++ src/main/java/insane96mcp/iguanatweaksreborn/module/items/misc/ItemDefinitionsReloadListener.java
@@ -17,14 +17,14 @@ import net.minecraftforge.event.TagsUpdatedEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.CopyOnWriteArrayList;
 
 @Mod.EventBusSubscriber(modid = InsaneSO.MOD_ID)
 public class ItemDefinitionsReloadListener extends SimpleJsonResourceReloadListener {
-	private static final List<ItemDefinition> DEFINITIONS = new ArrayList<>();
+	public static final List<ItemDefinition> DEFINITIONS = new CopyOnWriteArrayList<>();
 	public static final Map<Item, ItemDefinition.Durability> DURABILITY_MAP = new HashMap<>();
 	public static final Map<Item, Integer> ORIGINAL_DURABILITY = new HashMap<>();
 	public static final ItemDefinitionsReloadListener INSTANCE;
@@ -39,7 +39,7 @@ public class ItemDefinitionsReloadListener extends SimpleJsonResourceReloadListe
 
 	@Override
 	protected void apply(Map<ResourceLocation, JsonElement> map, ResourceManager resourceManager, ProfilerFiller profilerFiller) {
-		getDefinitions().clear();
+		DEFINITIONS.clear();
 		DURABILITY_MAP.clear();
 		for (var entry : map.entrySet()) {
 			try {
@@ -50,7 +50,7 @@ public class ItemDefinitionsReloadListener extends SimpleJsonResourceReloadListe
 
 				ItemDefinition itemDefinition = GSON.fromJson(entry.getValue(), ItemDefinition.class);
 				//itemStatistics.applyStats(false);
-				getDefinitions().add(itemDefinition);
+				DEFINITIONS.add(itemDefinition);
 			}
 			catch (JsonSyntaxException e) {
 				ISOLogHelper.error("Parsing error loading Item Definition %s: %s", entry.getKey(), e.getMessage());
@@ -60,7 +60,7 @@ public class ItemDefinitionsReloadListener extends SimpleJsonResourceReloadListe
 			}
 		}
 
-		ISOLogHelper.info("Loaded %s Item Definition", getDefinitions().size());
+		ISOLogHelper.info("Loaded %s Item Definition", DEFINITIONS.size());
 
 		/*for (var entry : Durability.entrySet()) {
 			entry.getValue().apply(entry.getKey());
@@ -87,10 +87,6 @@ public class ItemDefinitionsReloadListener extends SimpleJsonResourceReloadListe
 		for (var entry : DURABILITY_MAP.entrySet()) {
 			entry.getValue().apply(entry.getKey());
 		}
-	}
-
-	public synchronized static List<ItemDefinition> getDefinitions() {
-		return DEFINITIONS;
 	}
 
 	@Override